+++ src/user/user.resolver.ts
@@ -1,4 +1,6 @@
+import { UnauthorizedException } from '@nestjs/common';
 import { Args, ID, Mutation, Query, Resolver } from '@nestjs/graphql';
+import { GetCurrentUser } from '../common/decorators';
 import { UpdateUserDto } from './dto';
 import { User } from './schema/user.schema';
 import { UserService } from './user.service';
@@ -8,22 +10,47 @@ export class UserResolver {
     constructor(private readonly userService: UserService) {}
 
     @Query(() => [User])
-    findAll(): Promise<User[]> {
+    findAll(@GetCurrentUser('isAdmin') isAdmin: boolean): Promise<User[]> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('unauthorized');
+        }
+
         return this.userService.findAll();
     }
 
     @Query(() => User)
-    findOne(@Args('userId', { type: () => ID }) userId: string): Promise<User> {
+    findOne(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
+        @Args('userId', { type: () => ID }) userId: string
+    ): Promise<User> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('unauthorized');
+        }
+
         return this.userService.findOne(userId);
     }
 
     @Mutation(() => User)
-    updateUser(@Args('updateUserInput') dto: UpdateUserDto): Promise<User> {
+    updateUser(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
+        @Args('updateUserInput') dto: UpdateUserDto
+    ): Promise<User> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('unauthorized');
+        }
+
         return this.userService.update(dto);
     }
 
     @Mutation(() => User)
-    removeUser(@Args('userId', { type: () => ID }) userId: string): Promise<User> {
+    removeUser(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
+        @Args('userId', { type: () => ID }) userId: string
+    ): Promise<User> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('unauthorized');
+        }
+
         return this.userService.remove(userId);
     }
 }
